rename useData to useFetchData and move to own file
