+++ .storybook/stories/DataFiltering/lib/utils.ts
@@ -1,45 +1,4 @@
-import { useEffect, useState } from 'react';
-import {
-	getData,
-	GetDataParams,
-	GetDataPagination,
-	GetDataFilters,
-} from './getData';
-import { BusinessForAuditWithIndex } from './generateBusinessData';
-
-export type DashboardTableData = {
-	/** Whether the data is loading */
-	loading: boolean;
-	/** The loaded data */
-	data: BusinessForAuditWithIndex[];
-	/** The total number of pages in the search */
-	totalPages: number;
-	/** The total number of items found in the search */
-	totalItems: number;
-};
-
-export function useData({
-	sort,
-	filters,
-	pagination,
-}: GetDataParams): DashboardTableData {
-	const [loading, setLoading] = useState(false);
-	const [data, setData] = useState<BusinessForAuditWithIndex[]>([]);
-	const [totalPages, setTotalPages] = useState(0);
-	const [totalItems, setTotalItems] = useState(0);
-
-	useEffect(() => {
-		setLoading(true);
-		getData({ sort, filters, pagination }).then((response) => {
-			setData(response.data);
-			setTotalPages(response.totalPages);
-			setTotalItems(response.totalItems);
-			setLoading(false);
-		});
-	}, [sort, filters, pagination]);
-
-	return { loading, data, totalPages, totalItems };
-}
+import { GetDataPagination, GetDataFilters } from './getData';
 
 export const generateTableCaption = ({
 	loading,
